+++ src/Migration/Logging/Log/Builder/MigrationLogBuilder.php
@@ -136,6 +136,13 @@ class MigrationLogBuilder
     {
         \assert(\class_exists($logClass) && \is_subclass_of($logClass, AbstractMigrationLogEntry::class));
 
+        if ($this->exceptionTrace !== null) {
+            // remove args from trace, as they quickly become too large
+            foreach ($this->exceptionTrace as &$trace) {
+                unset($trace['args']);
+            }
+        }
+
         return new $logClass(
             $this->runId,
             $this->profileName,